+++ elf_examples/data_test.c
@@ -1,0 +1,6 @@
+static int counter = 10;
+
+int test_data(int amount) {
+    counter += amount;
+    return counter;
+}