+++ elf_examples/file_a.c
@@ -1,0 +1,3 @@
+int add_ten(int x) {
+    return x + 10;
+}