+++ elf_examples/file_b.c
@@ -1,0 +1,5 @@
+extern int add_ten(int x);
+
+int use_add_ten(int x) {
+    return add_ten(x) * 2;
+}