+++ elf_examples/rodata_test.c
@@ -1,0 +1,4 @@
+int test_rodata(int index) {
+    const char* msg = "Hello from rodata!";
+    return (int)msg[index];
+}